refactor: 충돌 해결 및 리뷰 타입 통합
