+++ apps/shop/app/review/my/page.tsx
@@ -1,9 +1,21 @@
 import { getUserReviews } from "@/src/features/myReviews/api/userReviewApi";
 import MyReviewsPage from "@/src/features/myReviews/components/MyReviewsPage";
 
-export default async function ReviewManagePage() {
+type PageProps = {
+    searchParams?: Promise<{
+        monthRange?: string | string[];
+        page?: string | string[];
+    }>;
+};
+
+export default async function ReviewManagePage({ searchParams }: PageProps) {
+    const parsedParams = await searchParams;
+
+    const monthRange = Number.parseInt(String(parsedParams?.monthRange));
+    const page = Number.parseInt(String(parsedParams?.page)) || 0;
+
     // 사용자의 리뷰 목록
-    const reviewsData = await getUserReviews(0, 10);
+    const reviewsData = await getUserReviews({ monthRange, page });
     const reviews = reviewsData.content || [];
 
     return (
@@ -14,4 +26,4 @@ export default async function ReviewManagePage() {
             currentPage={reviewsData.page || 0}
         />
     );
-} 
+}
